App/requirements.txt
# /requirements.txt

Flask
Flask-SQLAlchemy
Flask-Modals
pandas
pyodbc
numpy
openpyxl
wfastcgi
xlrd>=2.0.1
xlsxwriter
psutil
pytest
pandas_usaddress

# New libraries for doc generation:
python-docx==0.8.11
python-pptx==0.6.21
fpdf==1.7.2
